fix(client): keep the mobile pull-down chrome reachable in every canvas mode
On the touch layout the pull-down handle + chrome sheet (global controls +
the host row) lived INSIDE MobileTileView, which only renders in the
`workspace` canvas mode. So switching to a host that was still connecting/
warming — or a down/failed host — replaced the whole tile view with a
full-screen status canvas and took the pull-handle with it. On a still-
connecting host there's no 'switch back' button either, so a phone user was
stranded with no way to reach the host row and switch away.

Fix: extract the pull-handle + chrome drawer into MobilePullChrome and render
it in App.tsx as a sibling ABOVE the canvas <Switch> on the touch layout —
the exact placement the desktop ChromeBar already uses. The handle (and the
host row it opens) is now present in every mode: connecting, warming, down,
hostFailed, empty, and workspace.

- New MobilePullChrome (the pull-handle + top chrome drawer), self-contained.
- MobileTileView / CompactTileView drop the chrome + its now-unused
  status/appTitle/onOpenPalette props (they move to MobilePullChrome at the
  App shell level).
- Covering e2e: the host row stays reachable at the empty (non-workspace)
  canvas — the e2e-reachable proxy for the connecting/down class.
