+++ website/src/content/changelog/unreleased.mdx
@@ -120,6 +120,7 @@ version: Unreleased
 
 ### Fixed
 
+- <Change title="On a phone, you can no longer get stranded on a connecting or unreachable host" tags={["mobile", "multi-host"]}>Switching to a host that was still **connecting** (or one that's **down**) used to take the top **pull-down handle** away with it — the whole screen became the host's status card, and on a still-connecting host there was no "switch back" button, so a phone user was **stuck** with no way to reach the host row and pick another machine. The pull-down chrome now lives above the canvas (the mobile twin of the desktop's always-present host bar), so the handle — and the host row inside it — stays reachable in **every** state: connecting, unreachable, or even with no terminals open.</Change>
 - <Change title="Test and smoke runs no longer leak padi/kaval daemons" pr={1771} tags={["testing", "ci"]}>A daemon a test or smoke run spawns is deliberately **detached** so it can survive a server restart (the property some scenarios exercise) — but that meant it also outlived the **run** whenever teardown was skipped (a signal-killed run never runs its cleanup hooks) or scoped to the wrong workspace, and `ci/smoke.sh` reaped only its two kolu-servers, never their daemons — so every green smoke run left **2 padi + 2 kaval** behind. `@kolu/surface-daemon` now offers a third daemon lifetime, `boundToPid`: a test/smoke-spawned padi/kaval **watches the run that spawned it and exits cleanly the moment that run is gone**, so a killed or crashed run can't strand daemons. This is **internal only** — production daemons are untouched (they still live `forever`; the bind is opt-in and absent in production, so there is no way to weaken a real daemon), and no on-screen behavior changes. A pid-reuse inside the ~2s watch window is a documented residual (at worst one extra run's daemon lingers), not a leak class.</Change>
 - <Change title="Switching back to a host you've already visited is instant again" pr={1764} tags={["multi-host", "terminal", "canvas"]}>Returning to a machine you'd viewed earlier used to **rebuild the whole canvas** — the dock emptied and refilled, the terminal went black, and the Code tab dropped to **"Loading…"** for about a second — even though kolu still held all of that state from moments before. A switch-**back** is now instant: the dock shows the full terminal list in one frame, the panels stay put, and the Code tab keeps what it had, then a fresh copy loads the moment you're back. Each host you visit keeps its terminal list, saved session, activity feed, daemon status, and Code-tab reads **warm** until it leaves the pool — a host you're not looking at isn't polled, so nothing runs in the background; its state is simply held and refreshed the instant you return. The only thing you ever wait on is the terminal's own sub-second re-attach. First-time visits to a host you've never opened are unchanged.</Change>
 - <Change title="An HTML preview stops jumping to the top while you read it" pr={1755} tags={["code tab", "preview", "remote"]}>Reading a rendered HTML file in the Code tab, scrolled halfway down, the preview would **reload and snap back to the top** at seemingly random moments — even though the file's text never changed. The trigger was another terminal doing **git work in the same repo** (a branch checkout during a PR loop, most often on a remote host): a checkout re-writes working-tree files with the **same bytes but a new modification time**, and the preview keyed its cache-buster on that modification time — so an identical-content rewrite read as an edit and forced a reload. The preview now keys on a **hash of the file's actual bytes**: a real content change still reloads it live, but a same-content rewrite leaves your scroll position exactly where it was.</Change>
